requirements.txt
pandas
numpy
dash
dash-bootstrap-templates
dash-bootstrap-components
dash-extensions
boto3
pillow
gunicorn
